feat: prevent a account from being deleted by being associated with a registration

+++ src/main/java/br/edu/ifsp/spo/eventos/eventplatformbackend/account/AccountService.java
@@ -13,6 +13,7 @@ import br.edu.ifsp.spo.eventos.eventplatformbackend.common.exceptions.*;
 import br.edu.ifsp.spo.eventos.eventplatformbackend.common.recaptcha.RecaptchaService;
 import br.edu.ifsp.spo.eventos.eventplatformbackend.common.security.JwtService;
 import br.edu.ifsp.spo.eventos.eventplatformbackend.common.security.JwtUserDetails;
+import br.edu.ifsp.spo.eventos.eventplatformbackend.registration.RegistrationRepository;
 import com.auth0.jwt.interfaces.DecodedJWT;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -35,6 +36,7 @@ import java.util.UUID;
 @Slf4j
 public class AccountService {
     private final AccountRepository accountRepository;
+    private final RegistrationRepository registrationRepository;
     private final JwtService jwtService;
     private final PasswordEncoder passwordEncoder;
     private final RecaptchaService recaptchaService;
@@ -52,6 +54,12 @@ public class AccountService {
 
     public void delete(UUID accountId) {
         Account account = getAccount(accountId);
+
+//        Account com speaker não pq não é  not null a relação. Account registration sim
+
+        if(registrationRepository.existsByAccountId(accountId)) {
+            throw new BusinessRuleException(BusinessRuleType.ACCOUNT_DELETE_WITH_REGISTRATIONS);
+        }
         accountRepository.deleteById(accountId);
         log.info("Delete account id={}, name={}, email={}", account.getId(), account.getName(), account.getEmail());
     }
